corrected __sync__synchronise and error_check()

+++ src/simulator/emulatorinterface/communication/shm/SharedMem.java
@@ -49,20 +49,20 @@ public class SharedMem extends  IpcBase
 		return 0;
 	}
 	
-	public int update(int tidApp, int numReads){
+	public long update(int tidApp, int numReads){
 		get_lock(tidApp, shmAddress, COUNT);
-		int queue_size = SharedMem.shmreadvalue(tidApp, shmAddress, COUNT);
+		long queue_size = SharedMem.shmreadvalue(tidApp, shmAddress, COUNT);
 		queue_size -= numReads;
 
 		// update queue_size
 		SharedMem
-				.shmwrite(tidApp, shmAddress, COUNT, queue_size);
+				.shmwrite(tidApp, shmAddress, COUNT, (int)queue_size);
 		SharedMem.release_lock(tidApp, shmAddress, COUNT);
 		
 		return queue_size;
 	}
 	
-	public int totalProduced (int tidApp){
+	public long totalProduced (int tidApp){
 		return shmreadvalue(tidApp, shmAddress, COUNT + 4);
 	}
 	public void finish(){
@@ -93,7 +93,7 @@ public class SharedMem extends  IpcBase
 	// reads only the "value" from the packet struct. could be done using shmread() as well,
 	// but if we only need to read value this saves from the heavy JNI callback and thus saves
 	// on time.
-	native static int shmreadvalue(int tid, long pointer, int index);
+	native static long shmreadvalue(int tid, long pointer, int index);
 	
 	// write in the shared memory. needed in peterson locks.
 	native static int shmwrite(int tid,long pointer, int index, int val);
